+++ pydantic_encryption/__init__.py
@@ -1,1 +1,2 @@
 from .models import *
+from .annotations import *
